Move accessibility to new module
